feat: added compose and pipe functions
Added Compose4, Compose5, Pipe4, and Pipe5 functions to support function composition with up to five arguments.
Added ComposeN and PipeN functions to support function composition with an arbitrary number of functions that takes and return the same arguments.
Added ComposeUnsafe and PipeUnsafe functions to support function composition with an arbitrary number of functions taking any as arguments and return values.